backend/requirements.txt
setuptools>=70.0.0 
Django>=4.2.11 
djangorestframework>=3.14.0 
django-cors-headers>=4.4.1 
djangorestframework-simplejwt>=5.3.1 
psycopg2-binary>=2.9.10 
redis>=5.0.4 
channels>=4.1.0 
channels-redis>=4.2.0 
celery>=5.4.0 
Pillow>=10.4.0 
reportlab>=4.0.0 
drf-spectacular>=0.27.0 
drf-spectacular-sidecar>=2024.1.1 
python-dotenv==1.0.0 
pandas>=2.2.0 
numpy>=1.26.0 
scikit-learn>=1.4.0 
joblib==1.3.2 
gunicorn==21.2.0 
django-extensions>=3.2.3 
django-filter==23.5 
django-allauth==0.61.1 
django-import-export==3.3.5 
django-redis==5.4.0 
websockets==12.0 
django-model-utils==4.3.1 
django-storages==1.14.2 
boto3==1.34.14 
django-anymail==10.2 
django-ckeditor==6.7.0 
django-mptt==0.14.0 
django-taggit==4.0.0 
django-ratelimit==4.1.0 
django-axes==6.0.1 
pyotp==2.9.0 
qrcode==7.4.2 
